templates/python_restful_api/requirements.txt
aniso8601==8.0.0
boto3==1.9.197
botocore==1.12.229
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
marshmallow==3.2.0
marshmallow-sqlalchemy==0.17.0
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.3
Werkzeug==0.15.6
python-dotenv==0.10.3
gevent==1.4.0
packages/clv_common_py-2020.1.29.1-py3-none-any.whl